Kernels are now curried, in other words they should be created by passing in the parameter they will use to determine the dimension. It will then return a function that accepts two vectors with the dimension stored internally. This was done because the testing and training need to be done with the same dimension. It makes it easier to pass around a partially instantiated kernel that will store the dimension.
SVM now stores no state.  Since it is convenient to reuse the same object and because I worry about accidentally using stale state while testing I decided to just pass everything as needed when training. It will then pass back a "model" package.  It is one stop shopping for testing.  The model includes all the mistakes, the kernel used and the penalty.  To test, just pass the model back in with the testing data to the testWithModel method and it will take care of the rest.  Passing back the model is also convenient because it can be serialized and reused later.